size of ..

+++ 0x00-hello_world/6-size.c
@@ -11,11 +11,11 @@ char b;
 long int c;
 long long int d;
 float e;
-printf("Size of a char: %lu byte(s)\n", sizeof(b));
-printf("Size of int: %lu bytes\n", sizeof(a));
-printf("Size of a long int: %lu byte(s)\n", sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", sizeof(d));
-printf("Size of a float: %lu byte(s)\n", sizeof(e));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(b));
+printf("Size of int: %lu bytes\n", (unsigned long)sizeof(a));
+printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
 
 int64_t my_64bit_int;
 printf("Size of a char: %lu byte(s)\n", (2 * sizeof(b)));
